+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 1. Automatic disable thread support if `Threads` not found.
+2. Test case can be unregistered by `cutest_unregister_case()`.
 
 
 ## v3.0.3 (2024/04/23)